# docs/swagger.yaml
definitions:
  models.Interaction:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      type:
        $ref: '#/definitions/models.InteractionType'
      updated_at:
        type: string
      user_id:
        type: string
      video_id:
        type: string
    type: object
  models.InteractionType:
    enum:
    - like
    - view
    - comment
    type: string
    x-enum-varnames:
    - Like
    - View
    - Comment
  request.Interaction:
    properties:
      content:
        maxLength: 1000
        type: string
      type:
        allOf:
        - $ref: '#/definitions/models.InteractionType'
        enum:
        - like
        - comment
        - view
      user_id:
        type: string
      video_id:
        type: string
    required:
    - type
    - user_id
    - video_id
    type: object
...
info:
  contact: {}
paths:
  /interactions:
    get:
      consumes:
      - application/json
      description: Get a paginated list of all interactions
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Interaction'
            type: array
        "500":
          description: Internal server error
          schema:
            type: string
      summary: List all interactions
      tags:
      - interactions
    post:
      consumes:
      - application/json
      description: Create a new interaction between a user and a video
      parameters:
      - description: Interaction object
        in: body
        name: interaction
        required: true
        schema:
          $ref: '#/definitions/request.Interaction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Interaction'
        "400":
          description: Bad request
          schema:
            type: string
      summary: Create a new interaction
      tags:
      - interactions
  /interactions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing interaction
      parameters:
      - description: Interaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid interaction ID
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete an interaction
      tags:
      - interactions
    get:
      consumes:
      - application/json
      description: Get details of a specific interaction
      parameters:
      - description: Interaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/request.Interaction'
        "400":
          description: Invalid interaction ID
          schema:
            type: string
        "404":
          description: Interaction not found
          schema:
            type: string
      summary: Get an interaction by ID
      tags:
      - interactions
    put:
      consumes:
      - application/json
      description: Update an existing interaction
      parameters:
      - description: Interaction ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated interaction object
        in: body
        name: interaction
        required: true
        schema:
          $ref: '#/definitions/request.Interaction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/request.Interaction'
        "400":
          description: Invalid interaction ID or data
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Update an interaction
      tags:
      - interactions
...
  /users/{userID}/videos/{videoID}/interactions:
    get:
      consumes:
      - application/json
      description: Get all interactions for a specific user and video combination
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Video ID
        in: path
        name: videoID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Interaction'
            type: array
        "400":
          description: Invalid user or video ID
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get all interactions between a user and a video
      tags:
      - interactions
...
swagger: "2.0"